Ondalıklı Kaynaklar
Ondalıklı kaynaklar geri geldi. Bu daldan ondalıklı kaynaklarla yazmaya devam edeceğim.

+++ kod/başaşağıderebeyi/soyutkuruluş/ulus/Ulus.java
@@ -16,8 +16,8 @@ import java.util.List;
 
 public class Ulus {
 	public final Dünya dünya;
-	public final int[] envanter;
-	public final int[] gelir;
+	public final float[] envanter;
+	public final float[] gelir;
 	public final List<Şehir> şehirler;
 	public final List<Yol> yollar;
 	public final List<Takas> takaslar;
@@ -27,8 +27,8 @@ public class Ulus {
 
 	public Ulus(final Dünya dünya, final Color renk) {
 		this.dünya = dünya;
-		envanter = new int[DEĞERLER.length];
-		gelir = new int[DEĞERLER.length];
+		envanter = new float[DEĞERLER.length];
+		gelir = new float[DEĞERLER.length];
 		şehirler = new ArrayList<>();
 		yollar = new ArrayList<>();
 		takaslar = new ArrayList<>();
@@ -36,51 +36,51 @@ public class Ulus {
 		dünya.uluslar.add(this);
 	}
 	
-	public int durum(final Kaynak kaynak) {
+	public float durum(final Kaynak kaynak) {
 		return envanter[kaynak.sıra];
 	}
 	
-	public boolean dene(final Kaynak kaynak, final int miktar) {
+	public boolean dene(final Kaynak kaynak, final float miktar) {
 		return durum(kaynak) >= miktar;
 	}
 	
 	public boolean dene(final Kaynak kaynak) {
-		return dene(kaynak, 1);
+		return dene(kaynak, 1.0F);
 	}
 	
-	public boolean dene(final int[] miktarlar) {
+	public boolean dene(final float[] miktarlar) {
 		for (int i = 0; i < DEĞERLER.length; i++)
 			if (envanter[i] < miktarlar[i])
 				return false;
 		return true;
 	}
 	
-	public void ayarla(final Kaynak kaynak, final int miktar) {
+	public void ayarla(final Kaynak kaynak, final float miktar) {
 		envanter[kaynak.sıra] = miktar;
 	}
 	
-	public void ekle(final Kaynak kaynak, final int miktar) {
+	public void ekle(final Kaynak kaynak, final float miktar) {
 		ayarla(kaynak, durum(kaynak) + miktar);
 	}
 	
 	public void ekle(final Kaynak kaynak) {
-		ekle(kaynak, 1);
+		ekle(kaynak, 1.0F);
 	}
 	
-	public void ekle(final int[] miktarlar) {
+	public void ekle(final float[] miktarlar) {
 		for (int i = 0; i < DEĞERLER.length; i++)
 			envanter[i] += miktarlar[i];
 	}
 	
-	public void çıkar(final Kaynak kaynak, final int miktar) {
+	public void çıkar(final Kaynak kaynak, final float miktar) {
 		ekle(kaynak, -miktar);
 	}
 	
 	public void çıkar(final Kaynak kaynak) {
-		çıkar(kaynak, 1);
+		çıkar(kaynak, 1.0F);
 	}
 	
-	public void çıkar(final int[] miktarlar) {
+	public void çıkar(final float[] miktarlar) {
 		for (int i = 0; i < DEĞERLER.length; i++)
 			envanter[i] -= miktarlar[i];
 	}
@@ -102,7 +102,7 @@ public class Ulus {
 				gelir[i] += şehir.üretim[i];
 	}
 
-	public int gelir(final Kaynak kaynak) {
+	public float gelir(final Kaynak kaynak) {
 		return gelir[kaynak.sıra];
 	}
 }
